+++ src/test/java/com/iexec/sms/chain/Web3jServiceTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2023-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.iexec.sms.blockchain;
+package com.iexec.sms.chain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -28,7 +28,7 @@ import java.time.Duration;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @ExtendWith(SpringExtension.class)
-@EnableConfigurationProperties(value = BlockchainConfig.class)
+@EnableConfigurationProperties(value = ChainConfig.class)
 @TestPropertySource(properties = {
         "blockchain.id=134",
         "blockchain.is-sidechain=true",
@@ -39,19 +39,19 @@ import static org.assertj.core.api.Assertions.assertThat;
         "blockchain.gas-price-cap=22000000000" })
 class Web3jServiceTests {
     @Autowired
-    private BlockchainConfig blockchainConfig;
+    private ChainConfig chainConfig;
 
     @Test
-    void checkBlockchainConfig() {
-        BlockchainConfig expectedConfig = new BlockchainConfig(
+    void checkChainConfig() {
+        ChainConfig expectedConfig = new ChainConfig(
                 134, true, "https://bellecour.iex.ec",
                 "0x3eca1B216A7DF1C7689aEb259fFB83ADFB894E7f", Duration.ofSeconds(5),
                 1.0f, 22_000_000_000L
         );
-        assertThat(blockchainConfig).isEqualTo(expectedConfig);
+        assertThat(chainConfig).isEqualTo(expectedConfig);
     }
     @Test
     void shouldCreateInstance() {
-        assertThat(new Web3jService(blockchainConfig)).isNotNull();
+        assertThat(new Web3jService(chainConfig)).isNotNull();
     }
 }